Added a backlog log message that will add log messages to the logger when the backlog has filled up such that logs are now being discarded.
Motioned the buffer reset call to occur when pulling from the pool instead of when saving. THis is safer as it's a single point in the code instead of multiple points. The downside is, now buffers will have reset called for no reason, but the likelihood of re-using buffers is high, so this should add little overhead.

Refactored the code to never block log messages, even if full.

Added a serializer factory method to allow users of this library to specify their own serialization method for messages. The default is still JSON, but you can specify anything, now.

Added additional interface definitions to allow implementing structs a single place to define the basic interface for the Base log messages, which is too convenient to not use, generally.

Updated documentation to include serializerfactories & channel overflow information.